+++ ext/damerau_levenshtein/extconf.rb
@@ -1,8 +1,8 @@
 # Loads mkmf which is used to make makefiles for Ruby extensions
-require 'mkmf'
+require "mkmf"
 
 # The destination
-dir_config('damerau-levenshtein/damerau_levenshtein')
+dir_config("damerau-levenshtein/damerau_levenshtein")
 
 # Do the work
-create_makefile('damerau-levenshtein/damerau_levenshtein')
+create_makefile("damerau-levenshtein/damerau_levenshtein")