+++ lib/damerau-levenshtein.rb
@@ -1,8 +1,9 @@
 # encoding: UTF-8
 
-require 'damerau-levenshtein/version'
-require 'damerau-levenshtein/damerau_levenshtein'
+require "damerau-levenshtein/version"
+require "damerau-levenshtein/damerau_levenshtein"
 
+# Damerau-Levenshtein algorithm
 module DamerauLevenshtein
   extend DamerauLevenshteinBinding
 
@@ -11,6 +12,6 @@ module DamerauLevenshtein
   end
 
   def self.distance(str1, str2, block_size = 1, max_distance = 10)
-    distance_utf(str1.unpack('U*'), str2.unpack('U*'), block_size, max_distance)
+    distance_utf(str1.unpack("U*"), str2.unpack("U*"), block_size, max_distance)
   end
 end