+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+damerau-levenshtein CHANGELOG
+=============================
+
+1.1.0 -- relax and update gem dependencies, update specs
+
 1.0.3 -- gems update
 
 1.0.2 -- includes refactoring by https://github.com/luislavena to remove
-dependency on jeweler
+         dependency on jeweler
 
 1.0.1 -- fixed gem dependencies